Fix BF16 packing and bug with __bf16 casts.

+++ sbgemm_kernel_16x8_zvl256b.c
@@ -256,22 +256,22 @@ int CNAME(BLASLONG M, BLASLONG N, BLASLONG K, FLOAT alpha, IFLOAT *A, IFLOAT *B,
             BLASLONG bi = n_top * K;
 
             for (BLASLONG k=0; k<K; k++) {
-                result0+=(float)(A[ai+0]*B[bi+0]);
-                result1+=(float)(A[ai+1]*B[bi+0]);
-                result2+=(float)(A[ai+0]*B[bi+1]);
-                result3+=(float)(A[ai+1]*B[bi+1]);
-                result4+=(float)(A[ai+0]*B[bi+2]);
-                result5+=(float)(A[ai+1]*B[bi+2]);
-                result6+=(float)(A[ai+0]*B[bi+3]);
-                result7+=(float)(A[ai+1]*B[bi+3]);
-                result8+=(float)(A[ai+0]*B[bi+4]);
-                result9+=(float)(A[ai+1]*B[bi+4]);
-                result10+=(float)(A[ai+0]*B[bi+5]);
-                result11+=(float)(A[ai+1]*B[bi+5]);
-                result12+=(float)(A[ai+0]*B[bi+6]);
-                result13+=(float)(A[ai+1]*B[bi+6]);
-                result14+=(float)(A[ai+0]*B[bi+7]);
-                result15+=(float)(A[ai+1]*B[bi+7]);
+                result0+=(float)(A[ai+0])*(float)(B[bi+0]);
+                result1+=(float)(A[ai+1])*(float)(B[bi+0]);
+                result2+=(float)(A[ai+0])*(float)(B[bi+1]);
+                result3+=(float)(A[ai+1])*(float)(B[bi+1]);
+                result4+=(float)(A[ai+0])*(float)(B[bi+2]);
+                result5+=(float)(A[ai+1])*(float)(B[bi+2]);
+                result6+=(float)(A[ai+0])*(float)(B[bi+3]);
+                result7+=(float)(A[ai+1])*(float)(B[bi+3]);
+                result8+=(float)(A[ai+0])*(float)(B[bi+4]);
+                result9+=(float)(A[ai+1])*(float)(B[bi+4]);
+                result10+=(float)(A[ai+0])*(float)(B[bi+5]);
+                result11+=(float)(A[ai+1])*(float)(B[bi+5]);
+                result12+=(float)(A[ai+0])*(float)(B[bi+6]);
+                result13+=(float)(A[ai+1])*(float)(B[bi+6]);
+                result14+=(float)(A[ai+0])*(float)(B[bi+7]);
+                result15+=(float)(A[ai+1])*(float)(B[bi+7]);
                 ai+=2;
                 bi+=8;
             }
@@ -314,14 +314,14 @@ int CNAME(BLASLONG M, BLASLONG N, BLASLONG K, FLOAT alpha, IFLOAT *A, IFLOAT *B,
             BLASLONG bi = n_top * K;
 
             for (BLASLONG k=0; k<K; k++) {
-                result0+=(float)(A[ai+0]*B[bi+0]);
-                result1+=(float)(A[ai+0]*B[bi+1]);
-                result2+=(float)(A[ai+0]*B[bi+2]);
-                result3+=(float)(A[ai+0]*B[bi+3]);
-                result4+=(float)(A[ai+0]*B[bi+4]);
-                result5+=(float)(A[ai+0]*B[bi+5]);
-                result6+=(float)(A[ai+0]*B[bi+6]);
-                result7+=(float)(A[ai+0]*B[bi+7]);
+                result0+=(float)(A[ai+0])*(float)(B[bi+0]);
+                result1+=(float)(A[ai+0])*(float)(B[bi+1]);
+                result2+=(float)(A[ai+0])*(float)(B[bi+2]);
+                result3+=(float)(A[ai+0])*(float)(B[bi+3]);
+                result4+=(float)(A[ai+0])*(float)(B[bi+4]);
+                result5+=(float)(A[ai+0])*(float)(B[bi+5]);
+                result6+=(float)(A[ai+0])*(float)(B[bi+6]);
+                result7+=(float)(A[ai+0])*(float)(B[bi+7]);
                 ai+=1;
                 bi+=8;
             }
@@ -501,14 +501,14 @@ int CNAME(BLASLONG M, BLASLONG N, BLASLONG K, FLOAT alpha, IFLOAT *A, IFLOAT *B,
             BLASLONG bi = n_top * K;
 
             for (BLASLONG k=0; k<K; k++) {
-                result0+=(float)(A[ai+0]*B[bi+0]);
-                result1+=(float)(A[ai+1]*B[bi+0]);
-                result2+=(float)(A[ai+0]*B[bi+1]);
-                result3+=(float)(A[ai+1]*B[bi+1]);
-                result4+=(float)(A[ai+0]*B[bi+2]);
-                result5+=(float)(A[ai+1]*B[bi+2]);
-                result6+=(float)(A[ai+0]*B[bi+3]);
-                result7+=(float)(A[ai+1]*B[bi+3]);
+                result0+=(float)(A[ai+0])*(float)(B[bi+0]);
+                result1+=(float)(A[ai+1])*(float)(B[bi+0]);
+                result2+=(float)(A[ai+0])*(float)(B[bi+1]);
+                result3+=(float)(A[ai+1])*(float)(B[bi+1]);
+                result4+=(float)(A[ai+0])*(float)(B[bi+2]);
+                result5+=(float)(A[ai+1])*(float)(B[bi+2]);
+                result6+=(float)(A[ai+0])*(float)(B[bi+3]);
+                result7+=(float)(A[ai+1])*(float)(B[bi+3]);
                 ai+=2;
                 bi+=4;
             }
@@ -537,10 +537,10 @@ int CNAME(BLASLONG M, BLASLONG N, BLASLONG K, FLOAT alpha, IFLOAT *A, IFLOAT *B,
             BLASLONG bi = n_top * K;
 
             for (BLASLONG k=0; k<K; k++) {
-                result0+=(float)(A[ai+0]*B[bi+0]);
-                result1+=(float)(A[ai+0]*B[bi+1]);
-                result2+=(float)(A[ai+0]*B[bi+2]);
-                result3+=(float)(A[ai+0]*B[bi+3]);
+                result0+=(float)(A[ai+0])*(float)(B[bi+0]);
+                result1+=(float)(A[ai+0])*(float)(B[bi+1]);
+                result2+=(float)(A[ai+0])*(float)(B[bi+2]);
+                result3+=(float)(A[ai+0])*(float)(B[bi+3]);
                 ai+=1;
                 bi+=4;
             }
@@ -675,10 +675,10 @@ int CNAME(BLASLONG M, BLASLONG N, BLASLONG K, FLOAT alpha, IFLOAT *A, IFLOAT *B,
             BLASLONG bi = n_top * K;
 
             for (BLASLONG k=0; k<K; k++) {
-                result0+=(float)(A[ai+0]*B[bi+0]);
-                result1+=(float)(A[ai+1]*B[bi+0]);
-                result2+=(float)(A[ai+0]*B[bi+1]);
-                result3+=(float)(A[ai+1]*B[bi+1]);
+                result0+=(float)(A[ai+0])*(float)(B[bi+0]);
+                result1+=(float)(A[ai+1])*(float)(B[bi+0]);
+                result2+=(float)(A[ai+0])*(float)(B[bi+1]);
+                result3+=(float)(A[ai+1])*(float)(B[bi+1]);
                 ai+=2;
                 bi+=2;
             }
@@ -701,8 +701,8 @@ int CNAME(BLASLONG M, BLASLONG N, BLASLONG K, FLOAT alpha, IFLOAT *A, IFLOAT *B,
             BLASLONG bi = n_top * K;
 
             for (BLASLONG k=0; k<K; k++) {
-                result0+=(float)(A[ai+0]*B[bi+0]);
-                result1+=(float)(A[ai+0]*B[bi+1]);
+                result0+=(float)(A[ai+0])*(float)(B[bi+0]);
+                result1+=(float)(A[ai+0])*(float)(B[bi+1]);
                 ai+=1;
                 bi+=2;
             }
@@ -814,8 +814,8 @@ int CNAME(BLASLONG M, BLASLONG N, BLASLONG K, FLOAT alpha, IFLOAT *A, IFLOAT *B,
             BLASLONG bi = n_top * K;
 
             for (BLASLONG k=0; k<K; k++) {
-                result0+=(float)(A[ai+0]*B[bi+0]);
-                result1+=(float)(A[ai+1]*B[bi+0]);
+                result0+=(float)(A[ai+0])*(float)(B[bi+0]);
+                result1+=(float)(A[ai+1])*(float)(B[bi+0]);
                 ai+=2;
                 bi+=1;
             }
@@ -835,7 +835,7 @@ int CNAME(BLASLONG M, BLASLONG N, BLASLONG K, FLOAT alpha, IFLOAT *A, IFLOAT *B,
             BLASLONG bi = n_top * K;
 
             for (BLASLONG k=0; k<K; k++) {
-                result0+=(float)(A[ai+0]*B[bi+0]);
+                result0+=(float)(A[ai+0])*(float)(B[bi+0]);
                 ai+=1;
                 bi+=1;
             }
